+++ src/getpaid_core/__init__.py
@@ -1,6 +1,6 @@
 """Getpaid Core -- framework-agnostic payment processing."""
 
-__version__ = "3.1.0"
+__version__ = "3.2.0"
 
 from getpaid_core.enums import AutoName
 from getpaid_core.enums import BackendMethod
@@ -9,6 +9,7 @@ from getpaid_core.enums import FraudEvent
 from getpaid_core.enums import FraudStatus
 from getpaid_core.enums import PaymentEvent
 from getpaid_core.enums import PaymentStatus
+from getpaid_core.exceptions import BackendNotFoundError
 from getpaid_core.exceptions import ChargeFailure
 from getpaid_core.exceptions import CommunicationError
 from getpaid_core.exceptions import CredentialsError
@@ -16,6 +17,7 @@ from getpaid_core.exceptions import GetPaidException
 from getpaid_core.exceptions import InvalidCallbackError
 from getpaid_core.exceptions import InvalidTransitionError
 from getpaid_core.exceptions import LockFailure
+from getpaid_core.exceptions import ReconciliationRequiredError
 from getpaid_core.exceptions import RefundFailure
 from getpaid_core.flow import PaymentFlow
 from getpaid_core.processor import BaseProcessor
@@ -30,6 +32,7 @@ from getpaid_core.types import TransactionResult
 __all__ = [
     "AutoName",
     "BackendMethod",
+    "BackendNotFoundError",
     "BaseProcessor",
     "ChargeFailure",
     "ChargeResult",
@@ -47,6 +50,7 @@ __all__ = [
     "PaymentStatus",
     "PaymentUpdate",
     "PluginRegistry",
+    "ReconciliationRequiredError",
     "RefundFailure",
     "RefundResult",
     "TransactionResult",